@@ -3,32 +3,38 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@date-io/date-fns": "1.x",
-    "@fullcalendar/core": "^5.6.0",
-    "@fullcalendar/daygrid": "5.6.0",
-    "@fullcalendar/interaction": "5.6.0",
-    "@fullcalendar/react": "5.6.0",
-    "@fullcalendar/resource-timegrid": "^5.6.0",
-    "@fullcalendar/timegrid": "5.6.0",
-    "@material-ui/core": "^4.11.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.57",
+    "@date-io/date-fns": "2.15.0",
+    "@emotion/react": "^11.10.4",
+    "@emotion/styled": "^11.10.4",
+    "@fullcalendar/core": "^5.11.3",
+    "@fullcalendar/daygrid": "5.11.3",
+    "@fullcalendar/interaction": "5.11.3",
+    "@fullcalendar/react": "5.11.2",
+    "@fullcalendar/resource-timegrid": "^5.11.3",
+    "@fullcalendar/timegrid": "5.11.3",
+    "@material-ui/core": "^4.12.4",
+    "@material-ui/icons": "^4.11.3",
+    "@material-ui/lab": "^4.0.0-alpha.61",
     "@material-ui/pickers": "^3.3.10",
-    "clsx": "^1.1.1",
-    "date-fns": "^2.21.1",
-    "firebase": "^8.4.1",
-    "formik": "^2.2.6",
-    "material-table": "^1.69.2",
+    "@mui/lab": "^5.0.0-alpha.98",
+    "@mui/material": "^5.10.4",
+    "@mui/styles": "^5.10.3",
+    "@mui/x-date-pickers": "^5.0.1",
+    "clsx": "^1.2.1",
+    "date-fns": "^2.29.2",
+    "firebase": "^9.9.4",
+    "formik": "^2.2.9",
+    "material-table": "^2.0.3",
     "material-ui-color-picker": "^3.5.1",
-    "node-sass": "^4.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.3",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "redux": "^4.0.4",
+    "node-sass": "^7.0.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
+    "redux": "^4.2.0",
     "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.3.0",
+    "redux-thunk": "^2.4.1",
     "redux-zap": "^0.3.1"
   },
   "scripts": {
