@@ -1,5 +1,5 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 
 import { createStore, applyMiddleware } from "redux";
 import thunk from "redux-thunk";
@@ -7,6 +7,9 @@ import { Provider } from "react-redux";
 import { createLogger } from "redux-logger";
 import { MuiPickersUtilsProvider } from "@material-ui/pickers";
 import DateFnsUtils from "@date-io/date-fns";
+import { createTheme, ThemeProvider } from "@mui/material";
+import { AdapterDateFns } from "@mui/x-date-pickers/AdapterDateFns";
+import frLocale from "date-fns/locale/fr";
 
 import reducers from "./store";
 
@@ -14,6 +17,7 @@ import App from "./App";
 import "./index.scss";
 
 import * as serviceWorker from "./serviceWorker";
+import { LocalizationProvider } from "@mui/x-date-pickers";
 
 const middlewares = [thunk];
 
@@ -26,14 +30,22 @@ if (process.env.NODE_ENV === "development") {
 }
 
 const store = createStore(reducers, applyMiddleware(...middlewares));
+const theme = createTheme();
 
-ReactDOM.render(
+const root = createRoot(document.getElementById("root")); // createRoot(container!) if you use TypeScript
+root.render(
   <Provider store={store}>
-    <MuiPickersUtilsProvider utils={DateFnsUtils}>
-      <App />
-    </MuiPickersUtilsProvider>
-  </Provider>,
-  document.getElementById("root")
+    <ThemeProvider theme={theme}>
+      <LocalizationProvider
+        adapterLocale={frLocale}
+        dateAdapter={AdapterDateFns}
+      >
+        <MuiPickersUtilsProvider utils={DateFnsUtils}>
+          <App />
+        </MuiPickersUtilsProvider>
+      </LocalizationProvider>
+    </ThemeProvider>
+  </Provider>
 );
 
 // If you want your app to work offline and load faster, you can change
